backend/requirements.txt
anthropic
python-dotenv
httpx
fastapi
uvicorn[standard]
supabase
python-multipart
pydantic[email]
